+++ src/main/resources/mybatis-config.xml
@@ -15,7 +15,7 @@
         <typeAlias type="com.project.doubleshop.domain.category.entity.CategoryType" alias="CategoryType"/>
         <typeAlias type="com.project.doubleshop.domain.category.entity.DepthLevel" alias="DepthLevel"/>
         <typeAlias type="com.project.doubleshop.domain.common.mapper.param.RequestItemsWithCategory" alias="RequestItemsWithCategory"/>
-        <typeAlias type="com.project.doubleshop.domain.member.entity.v2.Member" alias="Member"/>
+        <typeAlias type="com.project.doubleshop.domain.member.entity.Member" alias="Member"/>
     </typeAliases>
     <typeHandlers>
         <typeHandler handler="com.project.doubleshop.domain.common.StatusTypeHandler" />
